+++ src/main/java/br/com/gjnv/petshop/repository/EnderecoRepository.java
@@ -4,9 +4,6 @@ import br.com.gjnv.petshop.model.Endereco;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import java.util.UUID;
-
 @Repository
-public interface EnderecoRepository extends JpaRepository<Endereco, Integer> {
+public interface EnderecoRepository extends JpaRepository<Endereco, Long> {
 }
-
